auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
pydantic[email]==2.6.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.7.0
redis>=6.2.0
python-jose[cryptography]
# google-cloud-pubsub==3.13.1
python-multipart  



